// web/src/api/generated/models/announcement-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface AnnouncementResponse {
    'id'?: number;
    'title'?: string;
    'content'?: string;
    'sort'?: number;
    'status'?: number;
    'statusDesc'?: string;
    'startTime'?: string;
    'endTime'?: string;
    'coverImage'?: string;
    'adminId'?: number;
    'viewCount'?: number;
    'readCount'?: number;
    'createTime'?: string;
    'updateTime'?: string;
}